+++ webserver/main/light_sensor.h
@@ -4,4 +4,4 @@
 #include <stdbool.h>
 
 void light_sensor_init(int i2c_bus_id);
-bool light_sensor_read(uint32_t *p_out);
+bool light_sensor_read(float *p_out);
